* Removed extract_esra.c and extract_esra_nothread.c (obsolete).
* Merged hcday_builder_mcclear_nothread.cxx into hcday_builder_mcclear.cxx.
This is implemented via a new option -nb-threads (also available in hcday_builder_esra).

* hcday_builder_esra and hcday_builder_mcclear compute now daily irradiation on a tilted plane.
This is implemented via a new option -split azim tilt albedo.
Note that hc3_thread_t::split() is a copy of compute_split() in extract.c.

* hcday_builder_esra and hcday_builder_mcclear compute also daily irradiation with the effect of relief shadows.
This is implemented via a new option -horizon.
THIS OPTION IS CURRENTLY VERY SLOW.

* Fixed memory leak in compute_srtm_horizon().
* Added compute_srtm_horizon_from_dataset(), a faster variant.



git-svn-id: http://svn.oie-lab.net/trunk/codes/solar_geometry@1374 67464621-8d75-4bde-9f6c-a8f136609eee